tracking: merge Python unit tests into test_unit.py.
